websocket: define CryptoPP::byte for older cryptopp
the "byte" typedef was moved into "CryptoPP" namespace in
https://github.com/weidai11/cryptopp/commit/00f9818b5d8e5aeb6c1057aa395317635bae07a3,
which was included by cryptopp since v6.0.0. but ubuntu/focal only shipped 5.6.4.
see https://packages.ubuntu.com/focal/libcrypto++-dev.

to ensure that seastar compiles on ubuntu/focal and with cryptopp < v6.0.0, we
need to define this typedef if it is missing.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>